[test] Split browser-specific code into browser_common.py. NFC
common.py was getting pretty huge and about 900 lines of it was
browser-test specific stuff.
